fix(audit): add trace message to failed logs
Also move `isAuthUser` to the Auth module as it made more sense there.
